+++ client/app/_layout.js
@@ -1,4 +1,4 @@
-import { Slot } from 'expo-router';
+import { Slot, SplashScreen } from 'expo-router';
 import { Platform } from 'react-native';
 import { Provider } from 'react-redux';
 import { PersistGate } from 'redux-persist/integration/react';
@@ -8,8 +8,22 @@ import { ThemeProvider } from '../context/theme';
 import Footer from '../components/footer/Footer';
 import { NetworkStatusProvider } from '../context/NetworkStatusProvider';
 import Header from '../screens/header';
+import { useEffect } from 'react';
+import { useFonts } from 'expo-font';
 
 export default function HomeLayout() {
+  const [loaded] = useFonts({
+    Inter: require('@tamagui/font-inter/otf/Inter-Medium.otf'),
+    InterBold: require('@tamagui/font-inter/otf/Inter-Bold.otf'),
+  });
+
+  useEffect(() => {
+    if (loaded) {
+      SplashScreen.hideAsync();
+    }
+  }, [loaded]);
+
+  if (!loaded) return null;
   return (
     <Provider store={store}>
       <PersistGate loading={null} persistor={persistor}>
