+++ client/packrat-ui/RText/index.js
@@ -1,0 +1,11 @@
+import { Text } from 'tamagui';
+
+const RText = ({ children, style, ...props }) => {
+  return (
+    <Text style={[style]} fontFamily="$body" {...props}>
+      {children}
+    </Text>
+  );
+};
+
+export default RText;
