--- client/packratui/RText/index.js
@@ -1,3 +1,0 @@
-import { Text } from 'tamagui';
-
-export default Text;
