+++ client/theme/tamagui.config.js
@@ -3,7 +3,7 @@ import { createInterFont } from '@tamagui/font-inter';
 import { createMedia } from '@tamagui/react-native-media-driver';
 import { shorthands } from '@tamagui/shorthands';
 import { themes, tokens } from '@tamagui/themes';
-import { createTamagui } from 'tamagui';
+import { createTamagui, createFont } from 'tamagui';
 
 const animations = createAnimations({
   bouncy: {
@@ -25,8 +25,40 @@ const animations = createAnimations({
   },
 });
 
-const headingFont = createInterFont();
-const bodyFont = createInterFont();
+// const headingFont = createInterFont();
+// const bodyFont = createInterFont();
+const interFont = createFont({
+  family: 'Inter, Helvetica, Arial, sans-serif',
+  // keys used for the objects you pass to `size`, `lineHeight`, `weight`
+  // and `letterSpacing` should be consistent. The `createFont` function
+  // will fill-in any missing values if `lineHeight`, `weight` or `letterSpacing`
+  // are subsets of `size`
+  size: {
+    1: 12,
+    2: 14,
+    3: 15,
+  },
+  lineHeight: {
+    // 1 will be 22
+    2: 22,
+  },
+  weight: {
+    1: '300',
+    // 2 will be 300
+    3: '600',
+  },
+  letterSpacing: {
+    1: 0,
+    2: -1,
+    // 3 will be -1
+  },
+  // (native) swap out fonts by face/style
+  face: {
+    300: { normal: 'InterLight', italic: 'InterItalic' },
+    600: { normal: 'InterBold' },
+  },
+});
+
 const config = createTamagui({
   animations,
   defaultTheme: 'dark',
@@ -34,8 +66,11 @@ const config = createTamagui({
   themeClassNameOnRoot: false,
   shorthands,
   fonts: {
-    heading: headingFont,
-    body: bodyFont,
+    // heading: headingFont,
+    // body: bodyFont,
+    inter: interFont,
+    heading: interFont,
+    body: interFont,
   },
   themes,
   tokens,
